fixed non-portable std::fabsf

+++ Proto-RPG/Sword.cpp
@@ -66,7 +66,7 @@ void Sword::update(float dt)
 	defaultCollider();
 	AnimatedSprite* animSprite = dynamic_cast<AnimatedSprite*>(_sprite);
 	// tparam in [0, 1], reaches 1 in the middle of animation and goes back to 0
-	float tparam = 1 - 2 * std::fabsf(animSprite->currentTime() - 0.5f);
+	float tparam = 1 - 2 * std::fabs(animSprite->currentTime() - 0.5f);
 	if (_facingDir == Direction::RIGHT)
 		_collider.adjust(0, std::min(0.1f + tparam, 0.9f), std::min(- 2 + tparam * 2, -0.2f), std::max(- tparam - 0.1f, -0.9f));
 	else if(_facingDir == Direction::LEFT)
